requirements.txt
anthropic>=0.40.0
fastapi>=0.111.0
feedparser>=6.0.11
groq>=0.9.0
python-dotenv>=1.0.0
requests>=2.32.0
uvicorn[standard]>=0.29.0
